Crate ad9959

Structs

Acr

Ad9959

A device driver for the AD9959 direct digital synthesis (DDS) chip.

Csr

Fr1

PartialAcr

PartialCsr

PartialFr1

ProfileSerializer

Represents a means of serializing a DDS profile for writing to a stream.

Enums

Address

Error

Possible errors generated by the AD9959 driver.

Mode

Indicates various communication modes of the DDS. The value of this enumeration is equivalent to the configuration bits of the DDS CSR register.

Traits

Interface

A trait that allows a HAL to provide a means of communicating with the AD9959.

Type Aliases

Channel