ad9959

Struct ProfileSerializer

pub struct ProfileSerializer { /* private fields */ }

Represents a means of serializing a DDS profile for writing to a stream.

Implementations

impl ProfileSerializer

pub fn new(mode: Mode) -> Self

Construct a new serializer.

Args

mode - The communication mode of the DDS.

pub fn push( &mut self, channels: Channel, ftw: Option<u32>, pow: Option<u14>, acr: Option<Acr>, )

Update a number of channels with the requested profile.

Args

channels - A set of channels to apply the configuration to.

ftw - If provided, indicates a frequency tuning word for the channels.

pow - If provided, indicates a phase offset word for the channels.

acr - If provided, indicates the amplitude control register for the channels. The ACR should be stored in the 3 LSB of the word. Note that if amplitude scaling is to be used, the “Amplitude multiplier enable” bit must be set.

pub fn finalize(&mut self) -> &[u32]

Get the serialized profile as a slice of 32-bit words.

Note

The serialized profile will be padded to the next 32-bit word boundary by adding dummy writes to the CSR or LSRR registers.

Returns

A slice of u32 words representing the serialized profile.