Crate ad9959

Structs

Ad9959

A device driver for the AD9959 direct digital synthesis (DDS) chip.

DdsConfig

The frozen DDS configuration.

ProfileSerializer

Represents a means of serializing a DDS profile for writing to a stream.

Enums

Channel

Specifies an output channel of the AD9959 DDS chip.

Error

Possible errors generated by the AD9959 driver.

Mode

Indicates various communication modes of the DDS. The value of this enumeration is equivalent to the configuration bits of the DDS CSR register.

Register

The configuration registers within the AD9959 DDS device. The values of each register are equivalent to the address.

Traits

Interface

A trait that allows a HAL to provide a means of communicating with the AD9959.